--- new_drum/task_templates/1_transforms/6_python3_image_transform/custom.py
@@ -1,129 +1,0 @@
-"""
-Copyright 2021 DataRobot, Inc. and its affiliates.
-All rights reserved.
-This is proprietary source code of DataRobot, Inc. and its affiliates.
-Released under the terms of DataRobot Tool and Utility Agreement.
-"""
-from typing import Union, Any
-import pandas as pd
-
-from datarobot_drum.custom_task_interfaces import TransformerInterface
-from .img_utils import (
-    b64_to_img,
-    img_to_b64,
-    img_to_grayscale,
-)
-
-
-def _process_image(raw_data: Union[str, bytes]) -> bytes:
-    img = b64_to_img(raw_data)
-    img = img_to_grayscale(img)
-    return img_to_b64(img)
-
-
-def fit(X, y, output_dir, **kwargs):
-    """ This hook defines how DataRobot will train this task. Even transform tasks need to be trained to learn/store information from training data
-    DataRobot runs this hook when the task is being trained inside a blueprint.
-    As an output, this hook is expected to create an artifact containing a trained object
-    The input parameters are passed by DataRobot based on project and blueprint configuration.
-
-    Parameters
-    -------
-    X: pd.DataFrame
-        Training data that DataRobot passes when this task is being trained.
-    y: pd.Series
-        Project's target column (None is passed for unsupervised projects).
-    output_dir: str
-        A path to the output folder; the artifact must be saved into this folder to be re-used in transform()
-
-    Returns
-    -------
-    None
-        fit() doesn't return anything, but must output an artifact (typically containing a trained object) into output_dir
-        so that the trained object can be used during scoring inside transform()
-    """
-    pass
-
-
-def load_model(code_dir: str) -> Any:
-    """
-    Can be used to load supported models if your model has multiple artifacts, or for loading
-    models that DRUM does not natively support
-
-    Parameters
-    ----------
-    code_dir : is the directory where model artifact and additional code are provided, passed in
-
-    Returns
-    -------
-    If used, this hook must return a non-None value
-    """
-    return _process_image
-
-
-def transform(data, transformer):
-    """This hook defines how DataRobot will use the trained object from fit() to transform new data.
-    DataRobot runs this hook when the task is used for scoring inside a blueprint.
-    As an output, this hook is expected to return the transformed data.
-    The input parameters are passed by DataRobot based on dataset and blueprint configuration.
-
-    Parameters
-    -------
-    data: pd.DataFrame
-        Data that DataRobot passes for transformation.
-    transformer: Any
-        Trained object, extracted by DataRobot from the artifact created inside fit().
-        In this example, it's a function
-
-    Returns
-    -------
-    pd.DataFrame
-        Returns a dataframe with transformed data.
-    """
-
-    return data.applymap(transformer)
-
-
-class CustomTask(TransformerInterface):
-    def fit(self, X, y, **kwargs):
-        """ This hook defines how DataRobot will train this task. Even transform tasks need to be trained to learn/store information from training data
-        DataRobot runs this hook when the task is being trained inside a blueprint.
-        As an output, this hook is expected to create an artifact containg a trained object [in this example - median of each numeric column], that is then used to transform new data.
-        The input parameters are passed by DataRobot based on project and blueprint configuration.
-
-        Parameters
-        -------
-        X: pd.DataFrame
-            Training data that DataRobot passes when this task is being trained.
-        y: pd.Series
-            Project's target column (None is passed for unsupervised projects).
-
-        Returns
-        -------
-        CustomTask
-            returns an object instance of class CustomTask that can be used in chained method calls
-        """
-
-        self.trn = MissingValuesMedianImputation(X)
-        self.trn.fit(X)
-
-        return self
-
-    def transform(self, data):
-        """ This hook defines how DataRobot will use the trained object from fit() to transform new data.
-        DataRobot runs this hook when the task is used for scoring inside a blueprint.
-        As an output, this hook is expected to return the transformed data.
-        The input parameters are passed by DataRobot based on dataset and blueprint configuration.
-
-        Parameters
-        -------
-        data: pd.DataFrame
-            Data that DataRobot passes for transformation.
-
-        Returns
-        -------
-        pd.DataFrame
-            Returns a dataframe with transformed data.
-        """
-
-        return self.trn.transform(data)
