+++ new_drum/task_templates/estimators/python_regression_with_custom_class/python_regression_with_custom_class.py
