+++ new_drum/task_templates/transforms/python3_image_transform/img_utils.py
