+++ new_drum/task_templates/transforms/python3_image_transform/requirements.txt
